smart-banana/requirements.txt
tensorflow>=2.13.0
streamlit>=1.25.0
flask>=2.3.0
flask-cors>=4.0.0
pillow>=9.5.0
matplotlib>=3.7.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
